+++ src/traversal/strategy/play_strategy.rs
@@ -5,25 +5,25 @@ use super::{strategy_trait::Strategy, training_strategy::{self, TrainingStrategy
 #[derive(Debug, Clone)]
 pub struct PlayStrategy {
     pub actions: usize,
-    pub play_strategy: [f32; DEFAULT_ACTION_COUNT],
+    pub play_strategy: [f64; DEFAULT_ACTION_COUNT],
 }
 
 impl Strategy for PlayStrategy {
     fn new(actions: usize) -> PlayStrategy {
         // println!("Strategy wasn't generated at train time, oh no!");
-        let mut play_strategy = [0f32; DEFAULT_ACTION_COUNT];
+        let mut play_strategy = [0f64; DEFAULT_ACTION_COUNT];
         for a in 0..actions {
-            play_strategy[a] = 1.0 / actions as f32;
+            play_strategy[a] = 1.0 / actions as f64;
         }
         play_strategy[0] = 1.0; // TODO - Assess if this at all likely
         PlayStrategy { actions: 0, play_strategy}
     }
 
-    fn get_current_strategy(&self, _iteration: usize) -> [f32; DEFAULT_ACTION_COUNT] {
+    fn get_current_strategy(&self, _iteration: usize) -> [f64; DEFAULT_ACTION_COUNT] {
         self.play_strategy
     }
 
-    fn from_existing_strategy(actions: usize, strategy: [f32; DEFAULT_ACTION_COUNT]) -> PlayStrategy {
+    fn from_existing_strategy(actions: usize, strategy: [f64; DEFAULT_ACTION_COUNT]) -> PlayStrategy {
         PlayStrategy { actions, play_strategy: strategy }
     }
 
@@ -35,7 +35,7 @@ impl Strategy for PlayStrategy {
 impl PlayStrategy {
     pub fn from_train_strategy(train_strategy: TrainingStrategy) -> PlayStrategy {       
         let mut normalizing_sum = 0.0;
-        let mut return_strategy = [0f32; DEFAULT_ACTION_COUNT];
+        let mut return_strategy = [0f64; DEFAULT_ACTION_COUNT];
 
         for r in 0..train_strategy.actions {
             normalizing_sum += train_strategy.strategy_sum[r];
@@ -48,13 +48,13 @@ impl PlayStrategy {
         } else {
             // If the normalizing sum is <= 0, then we have to assign equal probability to all actions
             for a in 0..train_strategy.actions {
-                return_strategy[a] = 1.0 / train_strategy.actions as f32;
+                return_strategy[a] = 1.0 / train_strategy.actions as f64;
             }
         };
         PlayStrategy { actions: train_strategy.actions, play_strategy: return_strategy }
     }
 
-    pub fn serialise(&self) -> Vec<f32> {
+    pub fn serialise(&self) -> Vec<f64> {
         self.play_strategy.to_vec()
     }
 }
