+++ src/traversal/strategy/training_strategy.rs
@@ -1,6 +1,6 @@
 use crate::config::*;
 
-use std::f32;
+use std::f64;
 
 use rand::Rng;
 
@@ -9,7 +9,7 @@ use crate::{thread_utils::with_rng, traversal::action_history::action::DEFAULT_A
 use super::strategy_trait::Strategy;
 
 // Provides an action index given the current strategy
-pub fn sample_strategy(strategy: &[f32], actions: usize) -> usize {
+pub fn sample_strategy(strategy: &[f64], actions: usize) -> usize {
     with_rng(|rng| {
         let mut r = rng.gen_range(0.0..1.0);
         for (i, &prob) in strategy.iter().enumerate() {
@@ -23,28 +23,28 @@ pub fn sample_strategy(strategy: &[f32], actions: usize) -> usize {
 }
 
 /// Constants for the strategy according to the Discounted CFR paper
-const ALPHA: f32 = 1.5;
-const BETA: f32 = 0.85;
-const GAMMA: f32 = 4.0;
+const ALPHA: f64 = 1.5;
+const BETA: f64 = 0.85;
+const GAMMA: f64 = 4.0;
 
 /// Strategy struct to hold the current strategy and the sum of all strategies
 #[derive(Clone, Debug)]
 pub struct TrainingStrategy {
     pub actions: usize,
-    pub strategy_sum: [f32; DEFAULT_ACTION_COUNT],
-    pub regrets_sum: [f32; DEFAULT_ACTION_COUNT],
+    pub strategy_sum: [f64; DEFAULT_ACTION_COUNT],
+    pub regrets_sum: [f64; DEFAULT_ACTION_COUNT],
 }
 
 impl Strategy for TrainingStrategy {
     fn new(actions: usize) -> TrainingStrategy {
         TrainingStrategy {
             actions,
-            strategy_sum: [0f32; DEFAULT_ACTION_COUNT],
-            regrets_sum: [0f32; DEFAULT_ACTION_COUNT],
+            strategy_sum: [0f64; DEFAULT_ACTION_COUNT],
+            regrets_sum: [0f64; DEFAULT_ACTION_COUNT],
         }
     }
 
-    fn get_current_strategy(&self, iteration: usize) -> [f32; DEFAULT_ACTION_COUNT] {
+    fn get_current_strategy(&self, iteration: usize) -> [f64; DEFAULT_ACTION_COUNT] {
         if iteration < MIN_SAMPLING_ITERATION_CUTOFF {
             self.threshold_strategy_update()
         } else {
@@ -52,11 +52,11 @@ impl Strategy for TrainingStrategy {
         }
     }
 
-    fn from_existing_strategy(actions: usize, strategy: [f32; DEFAULT_ACTION_COUNT]) -> Self {
+    fn from_existing_strategy(actions: usize, strategy: [f64; DEFAULT_ACTION_COUNT]) -> Self {
         TrainingStrategy {
             actions: actions,
             strategy_sum: strategy, // TODO - I did this when removing the current strategy ... not sure if correct
-            regrets_sum: [0f32; DEFAULT_ACTION_COUNT],
+            regrets_sum: [0f64; DEFAULT_ACTION_COUNT],
         }
     }
 
@@ -69,12 +69,12 @@ impl TrainingStrategy {
     // update the regrets
     pub fn update_strategy(
         &mut self,
-        strategy_utility: f32,
-        action_utilities: &Vec<f32>,
+        strategy_utility: f64,
+        action_utilities: &Vec<f64>,
         iteration: usize,
     ) {
         // println!("updating strategy with utilities, {} . {:?}", strategy_utility, action_utilities);
-        let iterf = iteration as f32;
+        let iterf = iteration as f64;
 
         // Temper the existing regrets sum according to Discount CFR
         let iter_coeff = iterf.powf(ALPHA);
@@ -103,7 +103,7 @@ impl TrainingStrategy {
     }
 
     // Updates the strategy sum based on the strategy calculated in the last iteration
-    fn update_strategy_sum_iter(&mut self, iteration: f32) {
+    fn update_strategy_sum_iter(&mut self, iteration: f64) {
         let current_strategy = self.get_current_strategy(iteration as usize);
         for index in 0..self.actions {
             // Add a DCRF weighted strategy to the strategy sum
@@ -113,7 +113,7 @@ impl TrainingStrategy {
     }
 
     // Used in later iterations to concentrate sample space
-    fn vanilla_strategy_update(&self) -> [f32; DEFAULT_ACTION_COUNT] {
+    fn vanilla_strategy_update(&self) -> [f64; DEFAULT_ACTION_COUNT] {
         let mut return_strategy = [0.0; DEFAULT_ACTION_COUNT];
         let mut normalizing_sum = 0.0;
         for r in 0..self.actions {
@@ -127,14 +127,14 @@ impl TrainingStrategy {
         } else {
             // If the normalizing sum is <= 0, then we have to assign equal probability to all actions
             for a in 0..self.actions {
-                return_strategy[a] = 1.0 / self.actions as f32;
+                return_strategy[a] = 1.0 / self.actions as f64;
             }
         };
         return_strategy
     }
 
     // Used in early exploration to ensure that all actions are sampled
-    fn threshold_strategy_update(&self) -> [f32; DEFAULT_ACTION_COUNT] {
+    fn threshold_strategy_update(&self) -> [f64; DEFAULT_ACTION_COUNT] {
         let mut return_strategy = [0.0; DEFAULT_ACTION_COUNT];
         let mut num_zero = 0;
         let mut normalizing_sum = 0.0;
@@ -146,17 +146,17 @@ impl TrainingStrategy {
             }
         }
 
-        let new_min = normalizing_sum / ((1.0 / MIN_SAMPLING_CUTOFF) - num_zero as f32);
-        normalizing_sum += num_zero as f32 * new_min;
+        let new_min = normalizing_sum / ((1.0 / MIN_SAMPLING_CUTOFF) - num_zero as f64);
+        normalizing_sum += num_zero as f64 * new_min;
 
         if normalizing_sum > 0.0 {
             for a in 0..self.actions {
-                return_strategy[a] = f32::max(self.regrets_sum[a], new_min) / normalizing_sum;
+                return_strategy[a] = f64::max(self.regrets_sum[a], new_min) / normalizing_sum;
             }
         } else {
             // If the normalizing sum is <= 0, then we have to assign equal probability to all actions
             for a in 0..self.actions {
-                return_strategy[a] = 1.0 / self.actions as f32;
+                return_strategy[a] = 1.0 / self.actions as f64;
             }
         };
         return_strategy
